bubbles-app vanilla JavaScript
AS part of this game there should be 4 circles and 4 arrows. On click of each circle the arrow should move towards the circle and once it hit the circle, the circle should change its color to grey. On click of the reset button it should move back to its original place.